fixup! TASK-83 - Moved file creation logic to a separate service
