+++ report/src/main/kotlin/ru/surf/report/controller/ReportListener.kt
@@ -1,39 +1,30 @@
 package ru.surf.report.controller
 
-import com.itextpdf.html2pdf.ConverterProperties
-import com.itextpdf.html2pdf.HtmlConverter
 import org.springframework.kafka.annotation.KafkaHandler
 import org.springframework.kafka.annotation.KafkaListener
 import org.springframework.kafka.annotation.RetryableTopic
 import org.springframework.messaging.handler.annotation.Payload
 import org.springframework.stereotype.Component
-import org.thymeleaf.context.Context
-import org.thymeleaf.spring6.SpringTemplateEngine
 import ru.surf.core.kafkaEvents.EndingEvent
 import ru.surf.report.service.ReportService
-import java.io.ByteArrayOutputStream
+import ru.surf.report.service.ReportWrapper
 
 
 @Component
 @KafkaListener(topics = ["core-topics"])
 class ReportListener(
-    private val templateEngine: SpringTemplateEngine,
     private val reportService: ReportService,
-    private val converterProperties: ConverterProperties,
+    private val reportWrapper: ReportWrapper
 ) {
 
     @KafkaHandler
     @RetryableTopic
     fun onEndingEvent(@Payload value: EndingEvent) {
-        val context = Context()
+        val report = reportWrapper.wrap(
+            reportService.getReport(value.eventId)
+        )
 
-        context.setVariable("report", reportService.getReport(value.eventId))
-        val html = templateEngine.process("report", context)
-
-        val target = ByteArrayOutputStream()
-        HtmlConverter.convertToPdf(html, target, converterProperties)
-
-        reportService.saveReport(target.toByteArray(), value.eventId)
+        reportService.saveReport(report, value.eventId)
     }
 
     @KafkaHandler(isDefault = true)
