+++ report/src/main/kotlin/ru/surf/report/service/ReportWrapper.kt
@@ -1,0 +1,7 @@
+package ru.surf.report.service
+
+import ru.surf.report.model.Report
+
+interface ReportWrapper {
+    fun wrap(report: Report): ByteArray
+}
